# Conflicts:
#	src/main/java/org/nicolas/controller/UserController.java
#	src/main/java/org/nicolas/model/Librarian.java
#	src/main/resources/LibraryDatabase/library.db

+++ src/main/java/org/nicolas/model/Librarian.java
@@ -2,74 +2,89 @@ package org.nicolas.model;
 
 import org.nicolas.database.LibraryDatabase;
 
-import java.io.Console;
+import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Objects;
 import java.util.ResourceBundle;
-import java.util.Scanner;
 import java.io.Console;
 
 public class Librarian extends User {
     private UserType userType;
+    private ResourceBundle messages;
 
-    //-------PROPOSING THIS CLASS AS THE ADMIN CLASS-------
-
+    //CONSTRUCTOR///////////////////////////////////////////////////////////////////////////////////////////////////////
     public Librarian(int userID, String name, String password) {
         super(userID, name, password);
         this.userType = UserType.LIBRARIAN;
     }
 
-    public void addBook (String ISBN, String title, String author, int copies, Console console, ResourceBundle messages) {
-        ArrayList<Book> booksFound = LibraryDatabase.getFilteredBooks(ISBN, title, author); //TODO check the return type and if it needs to be changed
+    //HELPER METHODS////////////////////////////////////////////////////////////////////////////////////////////////////
+    /**
+     * adds a book to the database
+     * @param ISBN the id of the book to add
+     * @param title the title of the book to add
+     * @param author the author of the book to add
+     * @param copies the number of copies of the new book
+     * @param console the system console
+     */
+    public void addBook (String ISBN, String title, String author, int copies, Console console) {
+        ArrayList<Book> booksFound = LibraryDatabase.getFilteredBooks(ISBN, title, author);
 
         if (booksFound.size() == 0 || booksFound.isEmpty()) {
             LibraryDatabase.insertIntoBooks(ISBN, title, author, copies);
-        } else {
-            System.out.println(messages.getString("book.exists"));
-            String ans = console.readLine().toUpperCase().charAt(0) + "";
+            return;
+        }
+
+        System.out.println(messages.getString("book.exists"));
+        String ans = console.readLine().toUpperCase().charAt(0) + "";
+
+        if (ans.equals("Y")) {
+            System.out.println(messages.getString("prompt.selectFromList"));
+            int existingBook = Integer.parseInt(console.readLine());
+            Book tempbook = booksFound.get(existingBook);
+            tempbook= LibraryDatabase.getBookThroughISBN(tempbook.getISBN());
+
+            System.out.printf("%s   %s, %s, COPIES : %d [BORROWED : %d    AVAILABLE : %d]\n",
+                    tempbook.getISBN(), tempbook.getTitle(), tempbook.getAuthor(), tempbook.getCopies(),
+                    tempbook.getBorrowedCopies(), tempbook.getAvailableCopies());
+
+            System.out.println(messages.getString("prompt.addCopies.question"));
+            ans = console.readLine().toUpperCase().charAt(0) + "";
 
             if (ans.equals("Y")) {
-                System.out.println(messages.getString("prompt.selectFromList"));
-                int existingBook = Integer.parseInt(console.readLine());
-                Book tempbook = booksFound.get(existingBook);
-                tempbook= LibraryDatabase.getBookThroughISBN(tempbook.getISBN());
-
-                System.out.printf("%s   %s, %s, COPIES : %d [BORROWED : %d    AVAILABLE : %d]\n",
-                        tempbook.getISBN(), tempbook.getTitle(), tempbook.getAuthor(), tempbook.getCopies(),
-                        tempbook.getBorrowedCopies(), tempbook.getAvailableCopies());
-
-                System.out.println(messages.getString("prompt.addCopies.question"));
-                ans = console.readLine().toUpperCase().charAt(0) + "";
-
-                if (ans.equals("Y")) {
-                    System.out.println(messages.getString("prompt.copies"));
-                    int copiesToAdd = Integer.parseInt(console.readLine());
-                    addCopiesToBook(tempbook, copiesToAdd);
-                    System.out.println(messages.getString("error.message.success"));
-                    return;
-                } else {
-                    System.out.println(messages.getString("error.message.operationAborted"));
-                    return;
-                }
-
-            } else if (ans.equals("N")) {
-                System.out.println(messages.getString("prompt.createNewBook"));
-                ans = console.readLine().toUpperCase().charAt(0) + "";
-
-                if (ans.equals("Y")) {
-                    LibraryDatabase.insertIntoBooks(ISBN, title, author, copies);
-                    System.out.println(messages.getString("error.message.success"));
-                    return;
-                } else {
-                    System.out.println(messages.getString("error.message.operationAborted"));
-                    return;
-                }
+                System.out.println(messages.getString("prompt.copies"));
+                int copiesToAdd = Integer.parseInt(console.readLine());
+                addCopiesToBook(tempbook, copiesToAdd);
+                System.out.println(messages.getString("error.message.success"));
+                return;
             } else {
-                System.out.println(messages.getString("error.message.invalidChoice"));
+                System.out.println(messages.getString("error.message.operationAborted"));
                 return;
             }
+
+        } else if (ans.equals("N")) {
+            System.out.println(messages.getString("prompt.createNewBook"));
+            ans = console.readLine().toUpperCase().charAt(0) + "";
+
+            if (ans.equals("Y")) {
+                LibraryDatabase.insertIntoBooks(ISBN, title, author, copies);
+                System.out.println(messages.getString("error.message.success"));
+                return;
+            } else {
+                System.out.println(messages.getString("error.message.operationAborted"));
+                return;
+            }
+        } else {
+            System.out.println(messages.getString("error.message.invalidChoice"));
+            return;
         }
     }
 
+    /**
+     * adds some new copies to a book and checks / changes the available copies count to match
+     * @param inputBook the book to be modified
+     * @param newCopiesNum the new number of copies
+     */
     public void addCopiesToBook (Book inputBook, int newCopiesNum) {
         int newTotalCopies = inputBook.getCopies() + newCopiesNum;
         inputBook.setCopies(newTotalCopies);
@@ -77,55 +92,31 @@ public class Librarian extends User {
         inputBook.setAvailableCopies(newAvCopies);
     }
 
+    /**
+     * changes the number of copies of a book in the database
+     * @param ISBN the book to update
+     * @param newNumOfCopies the new number of copies
+     */
     public void changeCopies (String ISBN, int newNumOfCopies) {
         LibraryDatabase.removeBookCopies(ISBN, newNumOfCopies);
     }
 
+    /**
+     * completly removes book from the database
+     * @param ISBN the id of the book to remove
+     */
     public void removeBook (String ISBN) {
         LibraryDatabase.deleteBookByIsbn(ISBN);
-//        Scanner console = new Scanner(System.in);
-//        ArrayList<Book> booksFound = LibraryDatabase.getFilteredBooks(ISBN, title, author); //TODO check the return type and if it needs to be changed
-//
-//        if (booksFound.size() == 0 || booksFound.isEmpty()) {
-//            System.out.println("No book to remove");
-//        } else  if (booksFound.size() > 1) {
-//            System.out.println("There seem to be multiple books matching your search. Select the one you wish to remove");
-//
-//            int existingBook = console.nextInt();
-//            Book tempbook = booksFound.get(existingBook);
-//            tempbook= LibraryDatabase.getBookThroughISBN(tempbook.getISBN());
-//
-//            System.out.printf("%s   %s, %s, COPIES : %d [BORROWED : %d    AVAILABLE : %d]",
-//                    tempbook.getISBN(), tempbook.getTitle(), tempbook.getAuthor(), tempbook.getCopies(),
-//                    tempbook.getBorrowedCopies(), tempbook.getAvailableCopies());
-//
-//            System.out.println("Do you wish to remove this book?");
-//            System.out.println("Y/N");
-//            ans = console.next().charAt(0);
-//
-//            if (ans == 'Y' || ans == 'y') {
-//                System.out.println("Please enter the number of copies to add : ");
-//                int copiesToAdd = console.nextInt();
-//                addCopiesToBook(tempbook, copiesToAdd);
-//            } else {
-//                System.out.println("Nothing added.");
-//            }
-//
-//        } else if (ans == 'N' || ans == 'n') {
-//            System.out.println("Create a new book?");
-//            System.out.println("Y/N");
-//            ans = console.next().charAt(0);
-//
-//            if (ans == 'Y' || ans == 'y') {
-//                LibraryDatabase.insertIntoBooks(ISBN, title, author, copies);
-//            } else {
-//                System.out.println("Nothing added.");
-//            }
-//        } else {
-//            System.out.println("Invalid choice"); //TODO create exception
-//        }
     }
 
+    //INHERITED H-METHODS///////////////////////////////////////////////////////////////////////////////////////////////
+
+    /**
+     * borrows a book for a student
+     * @param isbn the id of the book to borrow
+     * @param studentId the id of the user for which the book if borrowed
+     * @param console the system console
+     */
     @Override
     public void borrowBook(String isbn, int studentId, Console console) {
         Student tempStudent = LibraryDatabase.getStudentFromId(studentId);
@@ -160,12 +151,11 @@ public class Librarian extends User {
                 tempStudent.getBorrowedBooks().add(bookToBorrow);
 
                 // Reflect changes in the database
-                Date currentDate = new Date();
-                currentDate.getCurrentDate();
+                LocalDate currentDate = LocalDate.now();
                 //reflect change in the borrowing ->>>> check DB if needed
-                bookToBorrow.setAvailableCopies(bookToBorrow.getAvailableCopies() -1);
-                //Update the borrowed books count
                 bookToBorrow.setBorrowedCopies(bookToBorrow.getBorrowedCopies() + 1);
+                bookToBorrow.setAvailableCopies(bookToBorrow.getCopies() - bookToBorrow.getBorrowedCopies());
+                //Update the borrowed books count
 
                 LibraryDatabase.insertIntoBorrowedBooks(studentId, isbn, currentDate);
                 LibraryDatabase.updateBookCopies(bookToBorrow);
@@ -177,24 +167,88 @@ public class Librarian extends User {
                         bookToBorrow.getCopies(), bookToBorrow.getBorrowedCopies(), bookToBorrow.getAvailableCopies());
                 return;
             }
-            System.out.println("Book not available or no copies left. ");
+            System.out.println(messages.getString("book.unavailable"));
         }
     }
 
+    /**
+     * returns a book for a user and reflect the changes to the database
+     * @param isbn the id of the book to return
+     * @param userID the id of the user whishing to return the book
+     */
     public void returnBook(String isbn, int userID) {
+        Student tempStudent = LibraryDatabase.getStudentFromId(userID);
+        Book bookToReturn = null;
+
+        for (Book book : tempStudent.getBorrowedBooks()) { // Check if the student has this book borrowed
+            if (book.getISBN().equals(isbn)) {
+                bookToReturn = book;
+                break;
+            }
+        }
+
+        if (bookToReturn == null) {
+            System.out.println(messages.getString("book.not_borrowed"));
+            return;
+        }
+
+        tempStudent.getBorrowedBooks().remove(bookToReturn); // Remove from the list
+
+        // Update available/borrowed copies in the DB
+        bookToReturn.setAvailableCopies(bookToReturn.getAvailableCopies() + 1);
+        LibraryDatabase.updateBookCopies(bookToReturn);
+
+        // Remove from borrowedBooks table in DB
         LibraryDatabase.deleteFromBorrowedBooks(userID, isbn);
+
+        System.out.println(messages.getString("book.return.success") + bookToReturn.getTitle());
     }
 
+    /**
+     * adds a user to the database
+     * @param newStudentName the student's new name
+     * @param newStudentPassword the student's base password
+     */
     public void addUser (String newStudentName, String newStudentPassword) {
         LibraryDatabase.insertIntoUser(newStudentName, "STUDENT", newStudentPassword);
         LibraryDatabase.selectLastInsertedStudent();
     }
 
+    /**
+     * completely removes a user from the database
+     * @param userId the id of the user to remove
+     */
     public void removeUser (int userId) {
         LibraryDatabase.deleteStudentFromId(userId);
     }
 
+    //BASE METHODS//////////////////////////////////////////////////////////////////////////////////////////////////////
+    @Override
+    public String toString() {
+        return String.format("%d  %s  ROLE : %s\n", getUserID(), getName(), userType);
+    }
 
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) return false;
+        Librarian librarian = (Librarian) o;
+        return userType == librarian.userType && Objects.equals(messages, librarian.messages);
+    }
 
+    @Override
+    public int hashCode() {
+        return Objects.hash(userType, messages);
+    }
 
+    public UserType getUserType() {
+        return userType;
+    }
+
+    public void setUserType(UserType userType) {
+        this.userType = userType;
+    }
+
+    public ResourceBundle getMessages() {
+        return messages;
+    }
 }
